requirements.txt
ansimarkup==1.4.0
appdirs==1.4.3
atomicwrites==1.3.0
attrs==19.1.0
bandit==1.5.1
better-exceptions-fork==0.2.1.post6
black==19.3b0
blessings==1.7
bpython==0.17.1
certifi==2019.3.9
chardet==3.0.4
Click==7.0
colorama==0.4.1
coverage==4.5.2
curtsies==0.3.0
flake8==3.6.0
gitdb2==2.0.5
GitPython==2.1.11
greenlet==0.4.15
idna==2.8
loguru==0.2.5
mccabe==0.6.1
more-itertools==6.0.0
packaging==19.0
pbr==5.1.3
pluggy==0.9.0
py==1.8.0
pycodestyle==2.4.0
pyflakes==2.0.0
Pygments==2.3.1
pyparsing==2.3.1
pytest==4.3.0
pytest-sugar==0.9.2
python-dotenv==0.10.1
PyYAML==5.1
requests==2.21.0
six==1.12.0
smmap2==2.0.5
stevedore==1.30.1
termcolor==1.1.0
toml==0.10.0
typing==3.6.6
urllib3==1.24.1
wcwidth==0.1.7
